+++ src/AnomalyDetectors/LocalOutlierFactor.php
@@ -3,6 +3,7 @@
 namespace Rubix\ML\AnomalyDetectors;
 
 use Rubix\ML\Learner;
+use Rubix\ML\Scoring;
 use Rubix\ML\Ranking;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
@@ -42,7 +43,7 @@ use const Rubix\ML\EPSILON;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class LocalOutlierFactor implements Estimator, Learner, Ranking, Persistable, Stringable
+class LocalOutlierFactor implements Estimator, Learner, Scoring, Ranking, Persistable, Stringable
 {
     use PredictsSingle, ScoresSingle;