+++ src/Other/Traits/LoggerAware.php
@@ -42,4 +42,12 @@ trait LoggerAware
     {
         return $this->logger;
     }
+
+    /**
+     * Remove the logger and silence logging output.
+     */
+    public function silence() : void
+    {
+        $this->logger = null;
+    }
 }