+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+- 0.3.0
+    - Added Sentence tokenizer
+    - Can now silence Verbose learners with silence() method
+    - Added Scoring interface for estimators that score samples
+    - Deprecated the Ranking interface
+    
 - 0.2.2
     - Optimized CART quantile-based node splitting
     - Fixed CART and Extra Tree min purity increase post pruning
-    - Fixed ITree infinite loop same sample splitting
 
 - 0.2.1
     - Optimized Stop Word Filter